Address review comments.
Renamed controlCellular.py to control_cellular.py and added a class
called CellularModule to encapsulate the comms to the module.

Reformated the test-cellular-and-wifi-dut.py to have the private methods
last.

Changed the generic TestRunner class name to be more specific.
