+++ ci/lava/tests/test-cellular-and-wifi-dut.py
@@ -10,13 +10,89 @@ import re
 import random
 import string
 import time
-import controlCellular as cellular
+from control_cellular import CellularModule as cellular
 
 
-class TestRunner:
+class TestCellularAndWifiDut:
     """Class to encapsulate the testing of cellular connectivity on a DUT."""
 
     overall_result = "fail"
+
+    def test_cellular_setup(self, execute_helper):
+        """Test the cellular and wifi bevaviour."""
+        TestRunner.overall_result = "pass"
+
+        TestRunner._debug(self, execute_helper)
+
+    def test_cellular_initialise_cellular_module(self, execute_helper):
+        """Enable the cellular interface."""
+        assert TestRunner._control_cellular(self, enable=True)
+
+    def test_cellular_disable_wired_ethernet(self, execute_helper):
+        """Bring down the wired interface."""
+        assert TestRunner._disable_interface(self, "eth0", execute_helper)
+
+    def test_cellular_ping_test_wired_expect_fail(self, execute_helper):
+        """Check we cannot ping using the wired interface."""
+        assert TestRunner._run_ping_tests(self, "eth0", "fail", execute_helper)
+
+    def test_cellular_ping_test_cellular_expect_pass_test_1(
+        self, execute_helper
+    ):
+        """Check we can ping using the cellular interface."""
+        assert TestRunner._run_ping_tests(self, "usb0", "pass", execute_helper)
+
+    def test_cellular_disable_cellular(self, execute_helper):
+        """Bring down the wired interface."""
+        assert TestRunner._disable_interface(self, "usb0", execute_helper)
+
+    def test_cellular_wait_for_wlan(self, execute_helper):
+        """Wait for 60 seconds to give the wlan time to come up."""
+        time.sleep(30)
+        test_command = ["connmanctl", "scan", "wifi"]
+        err, stdout, stderr = execute_helper.execute_command(test_command)
+        time.sleep(30)
+        assert err == 0
+
+    def test_cellular_ping_test_wifi_expect_pass(self, execute_helper):
+        """Check we can ping using the wifi interface."""
+        assert TestRunner._run_ping_tests(
+            self, "wlan0", "pass", execute_helper
+        )
+
+    def test_cellular_ping_test_cellular_expect_fail(self, execute_helper):
+        """Check we cannot ping using the cellular interface."""
+        assert TestRunner._run_ping_tests(self, "usb0", "fail", execute_helper)
+
+    def test_cellular_reenable_cellular_module(self, execute_helper):
+        """Re-enable the cellular interface."""
+        assert TestRunner._enable_interface(self, "usb0", execute_helper)
+
+    def test_cellular_ping_test_wifi_expect_pass(self, execute_helper):
+        """Check we cannot ping using the wifi interface."""
+        assert TestRunner._run_ping_tests(
+            self, "wlan0", "fail", execute_helper
+        )
+
+    def test_cellular_ping_test_cellular_expect_pass_test_2(
+        self, execute_helper
+    ):
+        """Check we can ping using the cellular interface."""
+        assert TestRunner._run_ping_tests(self, "usb0", "pass", execute_helper)
+
+    def test_cellular_cleanup(self, execute_helper):
+        """Disable the cellular interface."""
+        # Bring down the cellular interface, to be clean.
+        TestRunner._control_cellular(self, enable=False)
+
+        # Bring up the wired interface.
+        TestRunner._enable_interface(self, "eth0", execute_helper)
+
+        assert True
+
+    def test_cellular_overall_result(self, execute_helper):
+        """Get overall result."""
+        assert TestRunner.overall_result
 
     def _run_ping_tests(self, interface, expected, execute_helper):
 
@@ -100,79 +176,3 @@ class TestRunner:
         debug_command = ["route"]
         err, stdout, stderr = execute_helper.execute_command(debug_command)
         print(stdout)
-
-    def test_cellular_setup(self, execute_helper):
-        """Test the cellular and wifi bevaviour."""
-        TestRunner.overall_result = "pass"
-
-        TestRunner._debug(self, execute_helper)
-
-    def test_cellular_initialise_cellular_module(self, execute_helper):
-        """Enable the cellular interface."""
-        assert TestRunner._control_cellular(self, enable=True)
-
-    def test_cellular_disable_wired_ethernet(self, execute_helper):
-        """Bring down the wired interface."""
-        assert TestRunner._disable_interface(self, "eth0", execute_helper)
-
-    def test_cellular_ping_test_wired_expect_fail(self, execute_helper):
-        """Check we cannot ping using the wired interface."""
-        assert TestRunner._run_ping_tests(self, "eth0", "fail", execute_helper)
-
-    def test_cellular_ping_test_cellular_expect_pass_test_1(
-        self, execute_helper
-    ):
-        """Check we can ping using the cellular interface."""
-        assert TestRunner._run_ping_tests(self, "usb0", "pass", execute_helper)
-
-    def test_cellular_disable_cellular(self, execute_helper):
-        """Bring down the wired interface."""
-        assert TestRunner._disable_interface(self, "usb0", execute_helper)
-
-    def test_cellular_wait_for_wlan(self, execute_helper):
-        """Wait for 60 seconds to give the wlan time to come up."""
-        time.sleep(30)
-        test_command = ["connmanctl", "scan", "wifi"]
-        err, stdout, stderr = execute_helper.execute_command(test_command)
-        time.sleep(30)
-        assert err == 0
-
-    def test_cellular_ping_test_wifi_expect_pass(self, execute_helper):
-        """Check we can ping using the wifi interface."""
-        assert TestRunner._run_ping_tests(
-            self, "wlan0", "pass", execute_helper
-        )
-
-    def test_cellular_ping_test_cellular_expect_fail(self, execute_helper):
-        """Check we cannot ping using the cellular interface."""
-        assert TestRunner._run_ping_tests(self, "usb0", "fail", execute_helper)
-
-    def test_cellular_reenable_cellular_module(self, execute_helper):
-        """Re-enable the cellular interface."""
-        assert TestRunner._enable_interface(self, "usb0", execute_helper)
-
-    def test_cellular_ping_test_wifi_expect_pass(self, execute_helper):
-        """Check we cannot ping using the wifi interface."""
-        assert TestRunner._run_ping_tests(
-            self, "wlan0", "fail", execute_helper
-        )
-
-    def test_cellular_ping_test_cellular_expect_pass_test_2(
-        self, execute_helper
-    ):
-        """Check we can ping using the cellular interface."""
-        assert TestRunner._run_ping_tests(self, "usb0", "pass", execute_helper)
-
-    def test_cellular_cleanup(self, execute_helper):
-        """Disable the cellular interface."""
-        # Bring down the cellular interface, to be clean.
-        TestRunner._control_cellular(self, enable=False)
-
-        # Bring up the wired interface.
-        TestRunner._enable_interface(self, "eth0", execute_helper)
-
-        assert True
-
-    def test_cellular_overall_result(self, execute_helper):
-        """Get overall result."""
-        assert TestRunner.overall_result
